+++ apps/backend/src/modules/activities/activity.entity.ts
@@ -1,5 +1,6 @@
 import { type Entity, type ValueOf } from "~/libs/types/types.js";
 import { type UserEntity } from "~/modules/users/user.entity.js";
+import { type UserSex } from "~/modules/users/users.js";
 
 import { type ActivityType } from "./libs/enums/enums.js";
 
@@ -9,6 +10,8 @@ class ActivityEntity implements Entity {
 	public commentCount: null | number;
 
 	public id: null | number;
+
+	public likesCount: null | number;
 
 	public payload: unknown;
 
@@ -24,6 +27,7 @@ class ActivityEntity implements Entity {
 		actionId,
 		commentCount,
 		id,
+		likesCount,
 		payload,
 		type,
 		updatedAt,
@@ -33,6 +37,7 @@ class ActivityEntity implements Entity {
 		actionId: number;
 		commentCount: null | number;
 		id: null | number;
+		likesCount: null | number;
 		payload: unknown;
 		type: ValueOf<typeof ActivityType>;
 		updatedAt: string;
@@ -41,6 +46,7 @@ class ActivityEntity implements Entity {
 	}) {
 		this.actionId = actionId;
 		this.id = id;
+		this.likesCount = likesCount;
 		this.payload = payload;
 		this.type = type;
 		this.updatedAt = updatedAt;
@@ -53,6 +59,7 @@ class ActivityEntity implements Entity {
 		actionId,
 		commentCount,
 		id,
+		likesCount,
 		payload,
 		type,
 		updatedAt,
@@ -62,6 +69,7 @@ class ActivityEntity implements Entity {
 		actionId: number;
 		commentCount: null | number;
 		id: number;
+		likesCount: null | number;
 		payload: unknown;
 		type: ValueOf<typeof ActivityType>;
 		updatedAt: string;
@@ -72,6 +80,7 @@ class ActivityEntity implements Entity {
 			actionId,
 			commentCount,
 			id,
+			likesCount,
 			payload,
 			type,
 			updatedAt,
@@ -95,6 +104,7 @@ class ActivityEntity implements Entity {
 			actionId,
 			commentCount: null,
 			id: null,
+			likesCount: null,
 			payload,
 			type,
 			updatedAt: "",
@@ -131,6 +141,7 @@ class ActivityEntity implements Entity {
 			id: number;
 			lastName: string;
 			nickname: null | string;
+			sex: ValueOf<typeof UserSex> | null;
 			updatedAt: string;
 		};
 		userId: number;
@@ -150,6 +161,7 @@ class ActivityEntity implements Entity {
 		actionId: number;
 		commentCount: null | number;
 		id: number;
+		likesCount: null | number;
 		payload: unknown;
 		type: ValueOf<typeof ActivityType>;
 		updatedAt: string;
@@ -161,6 +173,7 @@ class ActivityEntity implements Entity {
 			id: number;
 			lastName: string;
 			nickname: null | string;
+			sex: ValueOf<typeof UserSex> | null;
 			updatedAt: string;
 		};
 		userId: number;
@@ -169,6 +182,7 @@ class ActivityEntity implements Entity {
 			actionId: this.actionId,
 			commentCount: Number(this.commentCount),
 			id: this.id as number,
+			likesCount: Number(this.likesCount),
 			payload: this.payload,
 			type: this.type,
 			updatedAt: this.updatedAt,