+++ apps/backend/src/modules/activities/activity.repository.ts
@@ -3,6 +3,7 @@ import { type QueryBuilder } from "objection";
 import { SortOrder } from "~/libs/enums/enums.js";
 import { DatabaseTableName } from "~/libs/modules/database/database.js";
 import { type Repository, type ValueOf } from "~/libs/types/types.js";
+import { type ActivityLikeModel } from "~/modules/activity-likes/activity-likes.js";
 import { type CommentModel } from "~/modules/comments/comments.js";
 import { UserEntity } from "~/modules/users/user.entity.js";
 
@@ -26,6 +27,13 @@ class ActivityRepository implements Repository<ActivityEntity> {
 			.as("commentCount");
 	}
 
+	private getLikesCountQuery(): QueryBuilder<ActivityLikeModel> {
+		return this.activityModel
+			.relatedQuery<ActivityLikeModel>(RelationName.LIKES)
+			.count()
+			.as("likesCount");
+	}
+
 	public async create(activity: ActivityEntity): Promise<ActivityEntity> {
 		const activityModel = await this.activityModel
 			.query()
@@ -38,6 +46,7 @@ class ActivityRepository implements Repository<ActivityEntity> {
 			actionId: activityModel.actionId,
 			commentCount: EMPTY_COUNT,
 			id: activityModel.id,
+			likesCount: EMPTY_COUNT,
 			payload: activityModel.payload,
 			type: activityModel.type,
 			updatedAt: activityModel.updatedAt,
@@ -78,17 +87,20 @@ class ActivityRepository implements Repository<ActivityEntity> {
 		const activity = await this.activityModel
 			.query()
 			.findById(id)
+			.select(`${DatabaseTableName.ACTIVITIES}.*`, this.getLikesCountQuery())
+			.select(`${DatabaseTableName.ACTIVITIES}.*`, this.getCommentsCountQuery())
 			.withGraphJoined(
 				`${RelationName.USER}.${RelationName.USER_DETAILS}.${RelationName.AVATAR_FILE}`,
 			)
-			.castTo<ActivityModel | undefined>()
+			.castTo<(ActivityModel & ActivityCounts) | undefined>()
 			.execute();
 
 		return activity
 			? ActivityEntity.initialize({
 					actionId: activity.actionId,
-					commentCount: null,
+					commentCount: activity.commentCount,
 					id: activity.id,
+					likesCount: activity.likesCount,
 					payload: activity.payload,
 					type: activity.type,
 					updatedAt: activity.updatedAt,
@@ -102,6 +114,7 @@ class ActivityRepository implements Repository<ActivityEntity> {
 						nickname: activity.user.userDetails.nickname,
 						passwordHash: "",
 						passwordSalt: "",
+						sex: activity.user.userDetails.sex,
 						updatedAt: activity.user.updatedAt,
 					}),
 					userId: activity.userId,
@@ -112,6 +125,7 @@ class ActivityRepository implements Repository<ActivityEntity> {
 	public async findAll(userId: number): Promise<ActivityEntity[]> {
 		const activities = await this.activityModel
 			.query()
+			.select(`${DatabaseTableName.ACTIVITIES}.*`, this.getLikesCountQuery())
 			.select(`${DatabaseTableName.ACTIVITIES}.*`, this.getCommentsCountQuery())
 			.whereIn(
 				`${DatabaseTableName.ACTIVITIES}.userId`,
@@ -133,6 +147,7 @@ class ActivityRepository implements Repository<ActivityEntity> {
 				actionId: activity.actionId,
 				commentCount: activity.commentCount,
 				id: activity.id,
+				likesCount: activity.likesCount,
 				payload: activity.payload,
 				type: activity.type,
 				updatedAt: activity.updatedAt,
@@ -146,6 +161,7 @@ class ActivityRepository implements Repository<ActivityEntity> {
 					nickname: activity.user.userDetails.nickname,
 					passwordHash: "",
 					passwordSalt: "",
+					sex: activity.user.userDetails.sex,
 					updatedAt: activity.user.updatedAt,
 				}),
 				userId: activity.userId,
@@ -173,6 +189,7 @@ class ActivityRepository implements Repository<ActivityEntity> {
 					actionId: activity.actionId,
 					commentCount: null,
 					id: activity.id,
+					likesCount: null,
 					payload: activity.payload,
 					type: activity.type,
 					updatedAt: activity.updatedAt,
@@ -190,6 +207,7 @@ class ActivityRepository implements Repository<ActivityEntity> {
 			.query()
 			.findById(id)
 			.patch(activity.toNewObject())
+			.select(`${DatabaseTableName.ACTIVITIES}.*`, this.getLikesCountQuery())
 			.select(`${DatabaseTableName.ACTIVITIES}.*`, this.getCommentsCountQuery())
 			.returning("*")
 			.castTo<(ActivityModel & ActivityCounts) | undefined>()
@@ -200,6 +218,7 @@ class ActivityRepository implements Repository<ActivityEntity> {
 					actionId: updatedActivity.actionId,
 					commentCount: updatedActivity.commentCount,
 					id: updatedActivity.id,
+					likesCount: updatedActivity.likesCount,
 					payload: updatedActivity.payload,
 					type: updatedActivity.type,
 					updatedAt: updatedActivity.updatedAt,
@@ -213,6 +232,7 @@ class ActivityRepository implements Repository<ActivityEntity> {
 						nickname: updatedActivity.user.userDetails.nickname,
 						passwordHash: "",
 						passwordSalt: "",
+						sex: updatedActivity.user.userDetails.sex,
 						updatedAt: updatedActivity.user.updatedAt,
 					}),
 					userId: updatedActivity.userId,