+++ apps/backend/src/modules/friends/friend.repository.ts
@@ -49,6 +49,7 @@ class FriendRepository implements Repository<UserEntity> {
 			nickname: followingUser.userDetails.nickname,
 			passwordHash: followingUser.passwordHash,
 			passwordSalt: followingUser.passwordSalt,
+			sex: followingUser.userDetails.sex,
 			updatedAt: followingUser.updatedAt,
 		});
 	}
@@ -105,6 +106,7 @@ class FriendRepository implements Repository<UserEntity> {
 					nickname: user.userDetails.nickname,
 					passwordHash: user.passwordHash,
 					passwordSalt: user.passwordSalt,
+					sex: user.userDetails.sex,
 					updatedAt: user.updatedAt,
 				})
 			: null;
@@ -134,6 +136,7 @@ class FriendRepository implements Repository<UserEntity> {
 				nickname: user.userDetails.nickname,
 				passwordHash: user.passwordHash,
 				passwordSalt: user.passwordSalt,
+				sex: user.userDetails.sex,
 				updatedAt: user.updatedAt,
 			});
 		});
@@ -226,6 +229,7 @@ class FriendRepository implements Repository<UserEntity> {
 					nickname: user.userDetails.nickname,
 					passwordHash: user.passwordHash,
 					passwordSalt: user.passwordSalt,
+					sex: user.userDetails.sex,
 					updatedAt: user.updatedAt,
 				});
 			}),
@@ -270,6 +274,7 @@ class FriendRepository implements Repository<UserEntity> {
 					nickname: user.userDetails.nickname,
 					passwordHash: user.passwordHash,
 					passwordSalt: user.passwordSalt,
+					sex: user.userDetails.sex,
 					updatedAt: user.updatedAt,
 				});
 			}),
@@ -314,6 +319,7 @@ class FriendRepository implements Repository<UserEntity> {
 					nickname: user.userDetails.nickname,
 					passwordHash: user.passwordHash,
 					passwordSalt: user.passwordSalt,
+					sex: user.userDetails.sex,
 					updatedAt: user.updatedAt,
 				});
 			}),
@@ -339,6 +345,7 @@ class FriendRepository implements Repository<UserEntity> {
 			nickname: updatedSubscription.userDetails.nickname,
 			passwordHash: updatedSubscription.passwordHash,
 			passwordSalt: updatedSubscription.passwordSalt,
+			sex: updatedSubscription.userDetails.sex,
 			updatedAt: updatedSubscription.updatedAt,
 		});
 	}