+++ apps/backend/src/modules/users/user.controller.ts
@@ -122,6 +122,9 @@ class UserController extends BaseController {
 	 *                  type: string
 	 *                nickname:
 	 *                  type: string
+	 *                sex:
+	 *                  type: string
+	 *                  enum: [male, female, prefer-not-to-say]
 	 *      responses:
 	 *        '200':
 	 *          description: Successful operation