+++ apps/frontend/src/libs/hooks/use-handle-esc-press/use-handle-esc-press.hook.tsx
@@ -1,0 +1,23 @@
+import { useEffect } from "react";
+
+type Properties = {
+	onEscPress: () => void;
+};
+
+const useHandleEscPress = ({ onEscPress }: Properties): void => {
+	useEffect(() => {
+		const handleEscPress = (event: KeyboardEvent): void => {
+			if (event.key === "Escape") {
+				onEscPress();
+			}
+		};
+
+		document.addEventListener("keyup", handleEscPress);
+
+		return () => {
+			document.removeEventListener("keyup", handleEscPress);
+		};
+	}, [onEscPress]);
+};
+
+export { useHandleEscPress };