+++ apps/frontend/src/modules/activities/slices/activities.ts
@@ -1,6 +1,7 @@
 import {
 	createComment,
 	getAllCommentsToActivity,
+	likeActivity,
 	loadActivities,
 } from "./actions.js";
 import { actions } from "./activities.slice.js";
@@ -9,6 +10,7 @@ const allActions = {
 	...actions,
 	createComment,
 	getAllCommentsToActivity,
+	likeActivity,
 	loadActivities,
 };